feat(app): move app in app path
