+++ src/App.css
@@ -1,84 +1,11 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
 
-.App {
-    text-align: center;
-    height: 100%;
-}
-
-.App-logo {
-    height: 40vmin;
-    pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-    .App-logo {
-        animation: App-logo-spin infinite 20s linear;
-    }
-}
-
-.App-header {
-    background-color: #282c34;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-    color: white;
-}
-
-.App-link {
-    color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-    from {
-        transform: rotate(0deg);
-    }
-    to {
-        transform: rotate(360deg);
-    }
-}
-
-
-.Header {
-    font-family: 'Roboto', sans-serif;
-    margin-left: 2rem;
-    margin-right: 2rem;
-    width: calc(100% - 4rem);
-}
-
-.RedBack {
-    background-color: red;
-    color: white;
-    width: 40px;
-    border-radius: 5px;
-    height: 30px;
-    padding-left: 6px;
-    padding-right: 6px;
-    padding-bottom: 3px;
-    margin-left: 3px;
-}
-
 h2 {
     font-family: 'Roboto', sans-serif;
 }
 
 
-.contentContainer {
-    width: 100vw;
-    display: flex;
-    justify-content: center;
-}
 
-.contentWrapper {
-    width: 90vw;
-    max-width: 1400px;
-    min-height: 500px;
-    display: grid;
-    gap: 20px;
-    grid-template-columns: 1fr 6fr;
-}
 
 .close-x {
     stroke: black;
@@ -114,10 +41,6 @@ h2 {
         color: blue;
     }
 
-    .Diagrams {
-        width: 100%;
-    }
-
     .Menu {
         left: 1.5rem;
         top: 4rem
@@ -127,12 +50,6 @@ h2 {
 @media screen and (max-width: 799px) {
     body {
         overflow-x: hidden;
-    }
-
-    .Diagrams {
-        left: 0;
-        margin-left: 0.3rem;
-        margin-right: 0.3rem;
     }
 
     .Map {
@@ -152,12 +69,12 @@ h2 {
         margin: 0;
     }
 
-    .Header {
+    .logo {
         text-align: center;
         margin-bottom: 0;
     }
 
-    .Header > h1 {
+    .logo > h1 {
         margin-bottom: 0;
     }
