+++ src/app/index.ts
@@ -1,0 +1,6 @@
+import App from "./ui/app";
+import {WithRouter, WithContext, compose} from "./hocs";
+
+
+export default App;
+export {WithRouter, WithContext, compose};
