+++ src/components/logo/styles/logo.module.css
@@ -1,0 +1,18 @@
+.logo {
+    font-family: 'Roboto', sans-serif;
+    margin-left: 2rem;
+    margin-right: 2rem;
+    width: calc(100% - 4rem);
+}
+
+.redBack {
+    background-color: red;
+    color: white;
+    width: 40px;
+    border-radius: 5px;
+    height: 30px;
+    padding-left: 6px;
+    padding-right: 6px;
+    padding-bottom: 3px;
+    margin-left: 3px;
+}
